refactor: 필요없는 구문 제거 및 UseCase 패키지화
