+++ data/src/main/java/com/wakeup/data/repository/GlobeRepositoryImpl.kt
@@ -28,10 +28,6 @@ class GlobeRepositoryImpl @Inject constructor(
         globeLocalDataSource.deleteGlobe(globe.toEntity())
     }
 
-    override suspend fun getGlobeId(globeName: String): Long {
-        return globeLocalDataSource.getGlobeIdByName(globeName)
-    }
-
     override fun getGlobes(): Flow<List<Globe>> {
         return globeLocalDataSource.getGlobes().map {
             it.map { globeEntity -> globeEntity.toDomain() }
