+++ domain/src/main/java/com/wakeup/domain/repository/GlobeRepository.kt
@@ -13,8 +13,6 @@ interface GlobeRepository {
 
     suspend fun deleteGlobe(globe: Globe)
 
-    suspend fun getGlobeId(globeName: String): Long
-
     fun getGlobes(): Flow<List<Globe>>
 
     fun getMomentsByGlobe(globeId: Long): Flow<PagingData<Moment>>
