+++ domain/src/main/java/com/wakeup/domain/usecase/globe/CreateGlobeUseCase.kt
@@ -1,13 +1,13 @@
-package com.wakeup.domain.usecase
+package com.wakeup.domain.usecase.globe
 
 import com.wakeup.domain.model.Globe
 import com.wakeup.domain.repository.GlobeRepository
 import javax.inject.Inject
 
-class DeleteGlobeUseCase @Inject constructor(
+class CreateGlobeUseCase @Inject constructor(
     private val globeRepository: GlobeRepository
 ) {
     suspend operator fun invoke(globe: Globe) {
-        globeRepository.deleteGlobe(globe)
+        globeRepository.createGlobe(globe)
     }
 }
