+++ domain/src/main/java/com/wakeup/domain/usecase/globe/UpdateGlobeUseCase.kt
@@ -1,13 +1,13 @@
-package com.wakeup.domain.usecase
+package com.wakeup.domain.usecase.globe
 
 import com.wakeup.domain.model.Globe
 import com.wakeup.domain.repository.GlobeRepository
 import javax.inject.Inject
 
-class CreateGlobeUseCase @Inject constructor(
+class UpdateGlobeUseCase @Inject constructor(
     private val globeRepository: GlobeRepository
 ) {
     suspend operator fun invoke(globe: Globe) {
-        globeRepository.createGlobe(globe)
+        globeRepository.updateGlobe(globe)
     }
 }
