+++ domain/src/main/java/com/wakeup/domain/usecase/moment/SaveMomentUseCase.kt
@@ -1,7 +1,6 @@
-package com.wakeup.domain.usecase
+package com.wakeup.domain.usecase.moment
 
 import com.wakeup.domain.model.Moment
-import com.wakeup.domain.repository.GlobeRepository
 import com.wakeup.domain.repository.MomentRepository
 import com.wakeup.domain.repository.PictureRepository
 import com.wakeup.domain.repository.RelationRepository
@@ -10,11 +9,10 @@ import javax.inject.Inject
 class SaveMomentUseCase @Inject constructor(
     private val momentRepository: MomentRepository,
     private val pictureRepository: PictureRepository,
-    private val globeRepository: GlobeRepository,
     private val relationRepository: RelationRepository,
 ) {
     suspend operator fun invoke(moment: Moment) {
-        val globeId = globeRepository.getGlobeId(moment.globes.first().name)
+        val globeId = moment.globes.first().id
 
         if (moment.pictures.isEmpty()) {
             val momentId = momentRepository.saveMoment(moment)
