+++ presentation/src/main/java/com/wakeup/presentation/ui/addmoment/PlaceSearchViewModel.kt
@@ -2,7 +2,7 @@ package com.wakeup.presentation.ui.addmoment
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.wakeup.domain.usecase.SearchPlaceUseCase
+import com.wakeup.domain.usecase.place.SearchPlaceUseCase
 import com.wakeup.presentation.mapper.toPresentation
 import com.wakeup.presentation.model.PlaceModel
 import dagger.hilt.android.lifecycle.HiltViewModel
