+++ presentation/src/main/java/com/wakeup/presentation/ui/globe/GlobeViewModel.kt
@@ -2,8 +2,8 @@ package com.wakeup.presentation.ui.globe
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.wakeup.domain.usecase.CreateGlobeUseCase
-import com.wakeup.domain.usecase.GetGlobesUseCase
+import com.wakeup.domain.usecase.globe.CreateGlobeUseCase
+import com.wakeup.domain.usecase.globe.GetGlobesUseCase
 import com.wakeup.presentation.mapper.toDomain
 import com.wakeup.presentation.mapper.toPresentation
 import com.wakeup.presentation.model.GlobeModel
